fix: refactored config, separate config for destination and source for separate paramgen
