+++ source/config/config.go
@@ -1,0 +1,34 @@
+// Copyright © 2022 Meroxa, Inc. & Yalantis
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+//go:generate paramgen -output=paramgen.go Config
+
+package config
+
+import (
+	"github.com/conduitio-labs/conduit-connector-redshift/config"
+)
+
+// Config contains source-specific configurable values.
+type Config struct {
+	config.Configuration
+
+	// OrderingColumn is a name of a column that the connector will use for ordering rows.
+	OrderingColumn string `key:"orderingColumn" validate:"required,lowercase,excludesall= ,lte=127"`
+	// Snapshot is the configuration that determines whether the connector
+	// will take a snapshot of the entire table before starting cdc mode.
+	Snapshot bool `key:"snapshot" default:"true"`
+	// BatchSize is a size of rows batch.
+	BatchSize int `key:"batchSize" validate:"gte=1,lte=100000" default:"1000"`
+}
